The 'git commit -m' is used to commmit the files but we use 'git push' to make it live on the git hub like hosted platforms...
The '-m' in the 'git commit -m' refers to the message. The message may contain any unnecessary information or useful but the Messsage is mandatory. The maximum '-m's we can use is  2 i think, because when we create a new file in the repositry there are only two fields for the description section below the File and above the 'COMMIT' Option.